# backend/openapi.yaml
openapi: 3.1.0
info:
  title: KanziApp API
  version: "1.0.0"
  description: >
    Client-agnostic REST contract for the KanziApp backend (Spring Boot 4).
    This hand-authored spec is the source of truth (contract-first). JSON field
    naming is camelCase. All times are ISO-8601 UTC. Auth is a Bearer JWT access
    token; obtain one from /auth/login and refresh via /auth/refresh.
servers:
  - url: http://localhost:8080/api/v1
    description: Local dev
security:
  - bearerAuth: []
paths:
...
  /rooms:
    get:
      tags: [rooms]
      summary: Rooms the current user belongs to
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/RoomResponse' } } } } }
    post:
      tags: [rooms]
      summary: Create a room (caller becomes admin)
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateRoomRequest' } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/RoomResponse' } } } }
  /rooms/join:
    post:
      tags: [rooms]
      summary: Join a room by invite code
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/JoinRoomRequest' } } }
      responses:
        '200': { description: Joined, content: { application/json: { schema: { $ref: '#/components/schemas/RoomResponse' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /rooms/{roomId}:
    parameters: [ { $ref: '#/components/parameters/RoomId' } ]
    get:
      tags: [rooms]
      summary: A room the caller is a member of
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/RoomResponse' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
    patch:
      tags: [rooms]
      summary: Update room details (admin only)
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/UpdateRoomRequest' } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/RoomResponse' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
    delete:
      tags: [rooms]
      summary: Delete a room (admin only)
      responses:
        '204': { description: Deleted }
        '403': { $ref: '#/components/responses/Forbidden' }
  /rooms/{roomId}/membership:
    parameters: [ { $ref: '#/components/parameters/RoomId' } ]
    delete:
      tags: [rooms]
      summary: Leave the room (deletes the room if you are the sole member)
      responses:
        '204': { description: Left }
        '400': { $ref: '#/components/responses/BadRequest' }
  /rooms/{roomId}/members:
    parameters: [ { $ref: '#/components/parameters/RoomId' } ]
    get:
      tags: [rooms]
      summary: Members of a room
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/RoomMemberResponse' } } } } }
  /rooms/{roomId}/members/{userId}:
    parameters:
      - { $ref: '#/components/parameters/RoomId' }
      - { name: userId, in: path, required: true, schema: { type: string, format: uuid } }
    delete:
      tags: [rooms]
      summary: Remove a member (admin only; use /membership to leave)
      responses:
        '204': { description: Removed }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    RoomId: { name: roomId, in: path, required: true, schema: { type: string, format: uuid } }
  responses:
    BadRequest: { description: Validation or bad input, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
...
    Forbidden: { description: Not permitted (not a member / not admin / not owner), content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
    NotFound: { description: Resource not found, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
    Conflict: { description: Conflicts with existing data (duplicate), content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
  schemas:
    ApiError:
      type: object
      properties:
        status: { type: integer }
        error: { type: string }
        message: { type: string }
...
    CreateRoomRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, maxLength: 100 }
        description: { type: [string, "null"], maxLength: 500 }
    UpdateRoomRequest:
      type: object
      properties:
        name: { type: string, maxLength: 100 }
        description: { type: string, maxLength: 500 }
    JoinRoomRequest:
      type: object
      required: [code]
      properties:
        code: { type: string }
    RoomResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: [string, "null"] }
        code: { type: string }
        createdBy: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        memberCount: { type: integer }
        currentUserRole: { type: [string, "null"], enum: [admin, member, null] }
    RoomMemberResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        roomId: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        role: { type: string, enum: [admin, member] }
        joinedAt: { type: string, format: date-time }
        username: { type: [string, "null"] }
        displayName: { type: [string, "null"] }
        avatarUrl: { type: [string, "null"] }
...